api/pyproject.toml
[project]
name = "beyond-chatgpt"
version = "0.1.0"
description = "gpt-like-a-dev"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "openai>=1.77.0",
    "pydantic>=2.11.4",
    "python-multipart>=0.0.18",
    "PyPDF2>=3.0.1",
    "aiofiles>=23.2.1",
    "jupyter>=1.1.1",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "tavily-python>=0.5.0",
    "sentence-transformers>=2.2.0",
    "mangum>=0.17.0",
]
